Change back `items` to `controllers`
