+++ Module.php
@@ -3,11 +3,28 @@
 namespace mdm\admin;
 
 use Yii;
-use yii\helpers\ArrayHelper;
 use yii\helpers\Inflector;
 
 /**
- * Module
+ * GUI manager for RBAC.
+ * 
+ * Use `$controllerMap` to change property of controller. 
+ * To change listed menu, use property `$menus`.
+ * ```
+ * 'controllerMap' => [
+ *     'assignment' => [
+ *         'class' => 'mdm\admin\controllers\AssignmentController',
+ *         'userClassName' => 'app\models\User',
+ *         'idField' => 'user_id'
+ *     ]
+ * ],
+ * 'menus' => [
+ *     'assigment' => [
+ *         'label' => 'Grand Access' // change label
+ *     ],
+ *     'route' => null, // disable menu
+ * ],
+ * ```
  * 
  * @property string $mainLayout Main layout using for module.
  * Default to layout of parent module.
@@ -27,30 +44,11 @@ class Module extends \yii\base\Module
     public $defaultRoute = 'assigment';
 
     /**
-     * @var array
-     * List of extra controller of module.
-     * To disable core item, set visible to false.
-     * ```
-     * [
-     *     ...
-     *     'route' => [
-     *         'visible' => false
-     *     ],
-     *     'other-item' => [
-     *         'class' => 'path\to\OtherItemController'
-     *     ],
-     * ],
-     * ```
-     * @see [[\yii\base\Module::$controllerMap]]
-     */
-    public $items = [];
-
-    /**
+     * @var array 
      * @see [[items]]
-     * @var array
      */
     private $_menus;
-    
+
     /**
      * @inheritdoc
      */
@@ -60,71 +58,6 @@ class Module extends \yii\base\Module
      * @var string 
      */
     private $_mainLayout;
-
-    /**
-     * Core controller of moduls.
-     * @return array
-     */
-    protected function getCoreItems()
-    {
-        $config = components\Configs::instance();
-
-        return [
-            'assigment' => [
-                'class' => 'mdm\admin\items\AssigmentController'
-            ],
-            'role' => [
-                'class' => 'mdm\admin\items\RoleController'
-            ],
-            'permission' => [
-                'class' => 'mdm\admin\items\PermissionController'
-            ],
-            'route' => [
-                'class' => 'mdm\admin\items\RouteController'
-            ],
-            'rule' => [
-                'class' => 'mdm\admin\items\RuleController'
-            ],
-            'menu' => [
-                'class' => 'mdm\admin\items\MenuController',
-                'visible' => $config->db !== null && $config->db->schema->getTableSchema($config->menuTable) !== null
-            ],
-        ];
-    }
-
-    /**
-     * Normalize `controllerMap` before create controller.
-     */
-    private function normalizeController()
-    {
-        $controllers = [];
-        $this->_menus = [];
-        $mid = '/' . $this->getUniqueId() . '/';
-        $items = ArrayHelper::merge($this->getCoreItems(), $this->items);
-        foreach ($items as $id => $config) {
-            $label = Inflector::humanize($id);
-            $visible = true;
-            if (is_array($config)) {
-                $label = ArrayHelper::remove($config, 'label', $label);
-                $visible = ArrayHelper::remove($config, 'visible', true);
-            }
-            if ($visible) {
-                $this->_menus[] = ['label' => $label, 'url' => [$mid . $id]];
-                $controllers[$id] = $config;
-            }
-        }
-        $this->controllerMap = ArrayHelper::merge($this->controllerMap, $controllers);
-    }
-
-    /**
-     * @inheritdoc
-     */
-    public function createController($route)
-    {
-        $this->normalizeController();
-
-        return parent::createController($route);
-    }
 
     /**
      * Get main layout of 'left-menu' and 'right-menu' layout.
@@ -179,13 +112,57 @@ class Module extends \yii\base\Module
         }
         $this->_mainLayout = $path;
     }
-    
+
+    /**
+     * Get core menu
+     * @return array
+     */
+    protected function getCoreMenus()
+    {
+        $mid = '/' . $this->getUniqueId() . '/';
+        $ids = ['assignment', 'role', 'permission', 'route', 'rule', 'menu'];
+        $config = components\Configs::instance();
+        $result = [];
+        foreach ($ids as $id) {
+            if ($id !== 'menu' || ($config->db !== null && $config->db->schema->getTableSchema($config->menuTable) !== null)) {
+                $result[$id] = ['label' => Inflector::humanize($id), 'url' => [$mid . $id]];
+            }
+        }
+        foreach (array_keys($this->controllerMap) as $id) {
+            $result[$id] = ['label' => Inflector::humanize($id), 'url' => [$mid . $id]];
+        }
+        return $result;
+    }
+
     /**
      * Get avalible menu.
      * @return array
      */
     public function getMenus()
     {
+        if ($this->_menus === null) {
+            return $this->_menus = $this->getCoreMenus();
+        }
         return $this->_menus;
+    }
+
+    /**
+     * Set or add avalible menu.
+     * @param array $menus
+     */
+    public function setMenus($menus)
+    {
+        $mid = '/' . $this->getUniqueId() . '/';
+        $this->_menus = $this->getMenus();
+        foreach ($menus as $id => $value) {
+            if (empty($value)) {
+                unset($this->_menus[$id]);
+            } else {
+                $this->_menus[$id] = isset($this->_menus[$id]) ? array_merge($this->_menus[$id], $value) : $value;
+                if (!isset($this->_menus[$id]['url'])) {
+                    $this->_menus[$id]['url'] = [$mid . $id];
+                }
+            }
+        }
     }
 }
