+++ controllers/RoleController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace mdm\admin\items;
+namespace mdm\admin\controllers;
 
 use mdm\admin\models\AuthItem;
 use mdm\admin\models\searchs\AuthItem as AuthItemSearch;
@@ -10,7 +10,6 @@ use yii\filters\VerbFilter;
 use yii\rbac\Item;
 use Yii;
 use mdm\admin\components\MenuHelper;
-use yii\web\Response;
 use yii\helpers\Html;
 
 /**
@@ -19,7 +18,7 @@ use yii\helpers\Html;
  * @author Misbahul D Munir <misbahuldmunir@gmail.com>
  * @since 1.0
  */
-class PermissionController extends Controller
+class RoleController extends Controller
 {
 
     /**
@@ -43,8 +42,8 @@ class PermissionController extends Controller
      */
     public function actionIndex()
     {
-        $searchModel = new AuthItemSearch(['type' => Item::TYPE_PERMISSION]);
-        $dataProvider = $searchModel->search(Yii::$app->getRequest()->getQueryParams());
+        $searchModel = new AuthItemSearch(['type' => Item::TYPE_ROLE]);
+        $dataProvider = $searchModel->search(Yii::$app->request->getQueryParams());
 
         return $this->render('index', [
                 'dataProvider' => $dataProvider,
@@ -62,21 +61,32 @@ class PermissionController extends Controller
         $model = $this->findModel($id);
         $authManager = Yii::$app->getAuthManager();
         $avaliable = $assigned = [
+            'Roles' => [],
             'Permission' => [],
             'Routes' => [],
         ];
         $children = array_keys($authManager->getChildren($id));
         $children[] = $id;
+        foreach ($authManager->getRoles() as $name => $role) {
+            if (in_array($name, $children)) {
+                continue;
+            }
+            $avaliable['Roles'][$name] = $name;
+        }
         foreach ($authManager->getPermissions() as $name => $role) {
             if (in_array($name, $children)) {
                 continue;
             }
             $avaliable[$name[0] === '/' ? 'Routes' : 'Permission'][$name] = $name;
         }
-        foreach ($authManager->getChildren($id) as $name => $child) {
-            $assigned[$name[0] === '/' ? 'Routes' : 'Permission'][$name] = $name;
-        }
 
+        foreach ($authManager->getChildren($id) as $name => $child) {
+            if ($child->type == Item::TYPE_ROLE) {
+                $assigned['Roles'][$name] = $name;
+            } else {
+                $assigned[$name[0] === '/' ? 'Routes' : 'Permission'][$name] = $name;
+            }
+        }
         $avaliable = array_filter($avaliable);
         $assigned = array_filter($assigned);
 
@@ -91,7 +101,7 @@ class PermissionController extends Controller
     public function actionCreate()
     {
         $model = new AuthItem(null);
-        $model->type = Item::TYPE_PERMISSION;
+        $model->type = Item::TYPE_ROLE;
         if ($model->load(Yii::$app->getRequest()->post()) && $model->save()) {
             MenuHelper::invalidate();
 
@@ -139,29 +149,31 @@ class PermissionController extends Controller
         $post = Yii::$app->getRequest()->post();
         $roles = $post['roles'];
         $manager = Yii::$app->getAuthManager();
-        $parent = $manager->getPermission($id);
+        $parent = $manager->getRole($id);
         $error = [];
         if ($action == 'assign') {
             foreach ($roles as $role) {
-                $child = $manager->getPermission($role);
+                $child = $manager->getRole($role);
+                $child = $child ? : $manager->getPermission($role);
                 try {
                     $manager->addChild($parent, $child);
-                } catch (\Exception $exc) {
-                    $error[] = $exc->getMessage();
+                } catch (\Exception $e) {
+                    $error[] = $e->getMessage();
                 }
             }
         } else {
             foreach ($roles as $role) {
-                $child = $manager->getPermission($role);
+                $child = $manager->getRole($role);
+                $child = $child ? : $manager->getPermission($role);
                 try {
                     $manager->removeChild($parent, $child);
-                } catch (\Exception $exc) {
-                    $error[] = $exc->getMessage();
+                } catch (\Exception $e) {
+                    $error[] = $e->getMessage();
                 }
             }
         }
         MenuHelper::invalidate();
-        Yii::$app->getResponse()->format = Response::FORMAT_JSON;
+        Yii::$app->response->format = \yii\web\Response::FORMAT_JSON;
 
         return [$this->actionRoleSearch($id, 'avaliable', $post['search_av']),
             $this->actionRoleSearch($id, 'assigned', $post['search_asgn']),
@@ -171,13 +183,22 @@ class PermissionController extends Controller
     public function actionRoleSearch($id, $target, $term = '')
     {
         $result = [
+            'Roles' => [],
             'Permission' => [],
             'Routes' => [],
         ];
-        $authManager = Yii::$app->getAuthManager();
+        $authManager = Yii::$app->authManager;
         if ($target == 'avaliable') {
             $children = array_keys($authManager->getChildren($id));
             $children[] = $id;
+            foreach ($authManager->getRoles() as $name => $role) {
+                if (in_array($name, $children)) {
+                    continue;
+                }
+                if (empty($term) or strpos($name, $term) !== false) {
+                    $result['Roles'][$name] = $name;
+                }
+            }
             foreach ($authManager->getPermissions() as $name => $role) {
                 if (in_array($name, $children)) {
                     continue;
@@ -189,7 +210,11 @@ class PermissionController extends Controller
         } else {
             foreach ($authManager->getChildren($id) as $name => $child) {
                 if (empty($term) or strpos($name, $term) !== false) {
-                    $result[$name[0] === '/' ? 'Routes' : 'Permission'][$name] = $name;
+                    if ($child->type == Item::TYPE_ROLE) {
+                        $result['Roles'][$name] = $name;
+                    } else {
+                        $result[$name[0] === '/' ? 'Routes' : 'Permission'][$name] = $name;
+                    }
                 }
             }
         }
@@ -206,7 +231,7 @@ class PermissionController extends Controller
      */
     protected function findModel($id)
     {
-        $item = Yii::$app->getAuthManager()->getPermission($id);
+        $item = Yii::$app->getAuthManager()->getRole($id);
         if ($item) {
             return new AuthItem($item);
         } else {
